requirements.txt
dash>=2.14.0
dash-leaflet>=1.0.0
dash-bootstrap-components>=1.5.0
gunicorn>=21.2.0
pandas>=2.1.0
numpy>=1.26.0
xarray>=2023.8.0
requests>=2.31.0
shapely>=2.0.0
scipy>=1.11.0
geopandas>=0.14
rasterio>=1.3
Pillow>=10.0.0
matplotlib>=3.8.0
